% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{Artibeus_occ}
\alias{Artibeus_occ}
\title{Occurrence records for \emph{Artibeus} (example dataset)}
\format{
A data frame
}
\source{
GBIF
}
\usage{
Artibeus_occ
}
\description{
A dataset with occurrence records for \emph{Artibeus} used in vilma examples and
vignettes. Fill this in with the real description.
}
\keyword{datasets}
