% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{artibeus_tree}
\alias{artibeus_tree}
\title{Phylogeny for \emph{Artibeus} species}
\format{
An object of class \code{"phylo"} with X tips.
}
\source{
VerLife
}
\usage{
artibeus_tree
}
\description{
A phylogenetic tree used for examples in vilma.
}
\keyword{datasets}
