% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{car_tree}
\alias{car_tree}
\title{Example carnivore tree}
\format{
An object of class \code{"phylo"}.
}
\source{
VertLife
}
\usage{
car_tree
}
\description{
A phylogenetic tree for South American carnivores (example).
}
\keyword{datasets}
