% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{data_sa}
\alias{data_sa}
\title{Carnivores occurrences from South Africa}
\format{
A data frame
}
\source{
GBIF.
}
\usage{
data_sa
}
\description{
Occurrences from some terrestrial carnivore species from South Africa
}
\keyword{datasets}
