% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_dist.R
\name{example_dist}
\alias{example_dist}
\title{Example points for demonstration}
\usage{
example_dist()
}
\value{
An object of class \code{data.frame} or \code{matrix}.
}
\description{
Loads an example points (CSV format) included in the package.
}
\examples{
dist <- example_dist()
head(dist)
}
