% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_tree.R
\name{example_tree}
\alias{example_tree}
\title{Example phylogenetic tree for demonstration}
\usage{
example_tree()
}
\value{
An object of class \code{phylo}.
}
\description{
Loads an example tree (Newick format) included in the package.
}
\examples{
tree <- example_tree()
plot(tree)
}
