% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.root.node.R
\name{find.root.node}
\alias{find.root.node}
\title{Find the Root Node of a Rooted Phylogeny}
\usage{
find.root.node(tree)
}
\arguments{
\item{tree}{A phylogenetic tree. Must be an object of class \code{phylo}
from the \code{ape} package and must be rooted. The function will throw an
error if an unrooted tree is provided.}
}
\value{
An integer value corresponding to the node number of the root.
For a standard \code{phylo} object, this is typically the highest-numbered
internal node (e.g., for a tree with 5 tips, the root is node 6).
}
\description{
This function identifies the numerical identifier of the root node in a
phylogenetic tree of class \code{phylo} from the \code{ape} package.
It is a helper function for navigating the structure of \code{phylo} objects.
}
\examples{
# Create a simple rooted tree
library(ape)
tree <- read.tree(text = "((A:1, B:1):1, C:2);")
find.root.node(tree) # Should return the root node number (e.g., 5)

}
\seealso{
The function \code{link[ape]{is.rooted}}
}
