% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_kpam.R
\name{find_kpam}
\alias{find_kpam}
\title{Select optimal k for PAM clustering using silhouette width}
\usage{
find_kpam(pd.dis)
}
\arguments{
\item{pd.dis}{A dissimilarity object of class \code{\link[stats]{dist}}.
This must represent a valid pairwise distance matrix.}
}
\value{
A list with:
\describe{
\item{best_k}{Integer. The number of clusters that maximizes
the average silhouette width.}
\item{sil_val}{Data frame with columns:
\code{K} (evaluated cluster numbers) and
\code{avg_sil} (mean silhouette width for each k).}
}
}
\description{
Evaluates a range of cluster numbers (k) for Partitioning Around Medoids
(PAM) clustering based on average silhouette width and returns the
optimal number of clusters.
}
\details{
The function evaluates cluster numbers in the range \code{2:15}
(default fixed range) using \code{\link[cluster]{pam}} with
\code{diss = TRUE}. For each k, the mean silhouette width across clusters
is computed.

The optimal k is selected as the value that maximizes the average
silhouette width.
}
\examples{
\dontrun{
# pd.dis must be a 'dist' object
res <- find_kpam(pd.dis)
res$best_k
plot(res$sil_val$K, res$sil_val$avg_sil, type = "b")
}

}
\seealso{
\code{\link[cluster]{pam}}, \code{\link[cluster]{silhouette}}
}
