% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.vilma.cor.R
\name{plot.vilma.cor}
\alias{plot.vilma.cor}
\title{Plot VILMA Correlation Results}
\usage{
\method{plot}{vilma.cor}(x, ...)
}
\arguments{
\item{x}{An object of class \code{vilma.cor}.}

\item{...}{Additional arguments (currently not used).}
}
\value{
A plot is produced. Invisibly returns \code{x}.
}
\description{
Provides a quick visualization of a \code{vilma.cor} object by plotting the
relationship between species richness and phylogenetic diversity (PD) values.
The method automatically selects the best-supported model based on the minimum
AIC value stored in \code{x$AIC}. If \pkg{ggplot2} is available, a scatterplot
with a fitted trend line is produced; otherwise, the function falls back to a
base R plot using the selected fitted model object.
}
\details{
Model selection is performed by choosing the model name corresponding to the
smallest value in \code{x$AIC}. When \pkg{ggplot2} is installed, the function
plots \code{x$Data} using \code{sr.vals} on the x-axis and \code{pd.vals} on the
y-axis, and overlays a fitted smooth with \code{geom_smooth(method = <best model>)}.
If \pkg{ggplot2} is not installed, the function plots the best model stored in
\code{x$Models[[method]]} using base R.
}
\seealso{
\code{\link{vilma.cor}}
}
\author{
Omar Daniel Leon-Alvarado \url{https://leon-alvarado.weebly.com/}
J. Angel Soto-Centeno \url{https://www.mormoops.com/}
}
