% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.vilma.region.R
\name{plot.vilma.region}
\alias{plot.vilma.region}
\title{Plot VILMA Regionalization Rasters}
\usage{
\method{plot}{vilma.region}(x, ...)
}
\arguments{
\item{x}{An object of class \code{vilma.region}, typically generated by
\code{\link{vilma.regionalize}}.}

\item{...}{Additional arguments (currently not used).}
}
\value{
Produces raster plots for the region IDs and mean within-region
dissimilarity. Invisibly returns \code{x}.
}
\description{
Provides a quick visualization of a \code{vilma.region} object produced by
\code{\link{vilma.regionalize}}. The plot shows two maps in a vertical layout:
(i) the raster of region (group) IDs and (ii) the raster of mean within-region
phylogenetic dissimilarity (mean PD distance) per cell.
}
\details{
This method is intended for quick inspection of regionalization results.
Internally it uses \code{\link[terra]{plot}} to display the rasters stored in
\code{x$raster}. A small helper is included to optionally label cells with
numeric values when the raster contains finite values.
}
\seealso{
\code{\link{vilma.regionalize}}, \code{\link[terra]{plot}}
}
\author{
Omar Daniel Leon-Alvarado \url{https://leon-alvarado.weebly.com/}
J. Angel Soto-Centeno \url{https://www.mormoops.com/}
}
