% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.vilma.dist.R
\name{print.vilma.dist}
\alias{print.vilma.dist}
\title{Print Summary of a VILMA Distribution Object}
\usage{
\method{print}{vilma.dist}(x, ...)
}
\arguments{
\item{x}{An object of class \code{vilma.dist}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Provides a concise textual summary of a \code{vilma.dist} object,
including the number of species, number of cells, total records, and basic species abundance statistics.
}
\details{
This function is designed to give a quick overview of the distribution data stored in a VILMA object.
It prints the number of unique taxa, spatial cells, total records, and summary statistics for species abundance.
}
\author{
Omar Daniel Leon-Alvarado \url{https://leon-alvarado.weebly.com/}
J. Angel Soto-Centeno \url{https://www.mormoops.com}
}
