% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.vilma.null.R
\name{print.vilma.null}
\alias{print.vilma.null}
\title{Print Summary of a VILMA Null Model}
\usage{
\method{print}{vilma.null}(x, ...)
}
\arguments{
\item{x}{An object of class \code{vilma.null}, typically generated by a VILMA null model function.}

\item{...}{Additional arguments (currently unused).}
}
\value{
Invisibly returns the input object. Primary purpose is printing a summary to the console.
}
\description{
Displays a textual summary of a \code{vilma.null} object. Depending on the null model method,
it prints either global PD statistics or cell-specific SES and p-values.
}
\details{
For objects with \code{Method = "global"}, the function reports observed PD, mean and SD of null PD,
standardized effect size (SES), and whether the p-value is \code{< 0.05}. For \code{Method = "cell"}, it
prints summary statistics of PD, null PD, SES per cell, and the number of cells with significant p-values.
}
\author{
Omar Daniel Leon-Alvarado \url{https://leon-alvarado.weebly.com/}
J. Angel Soto-Centeno \url{https://www.mormoops.com}
}
