% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.vilma.pd.R
\name{print.vilma.pd}
\alias{print.vilma.pd}
\title{Print Summary of a VILMA Phylogenetic Diversity Object}
\usage{
\method{print}{vilma.pd}(x, ...)
}
\arguments{
\item{x}{An object of class \code{vilma.pd}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Provides a concise textual summary of a \code{vilma.pd} object, including:
the number of species, number of spatial cells, total records, species abundance statistics,
and summary statistics of phylogenetic diversity (PD) metrics.
}
\details{
This function prints a quick overview of the phylogenetic diversity stored in a VILMA object.
It summarizes the distribution data and PD values per cell for quick inspection.
}
\author{
Omar Daniel Leon-Alvarado \url{https://leon-alvarado.weebly.com/}
J. Angel Soto-Centeno \url{https://www.mormoops.com}
}
