% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{sa_pol}
\alias{sa_pol}
\title{South Africa polygon}
\format{
An \code{sf} object.
}
\source{
DIVA-GIS.
}
\usage{
sa_pol
}
\description{
Spatial polygon for South Africa used for mapping examples.
}
\keyword{datasets}
